# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/properties": {
      "get": {
        "operationId": "Properties_list",
        "parameters": [
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "name": {
                            "type": "string"
                          },
                          "addressLine1": {
                            "type": "string"
                          },
                          "addressLine2": {
                            "nullable": true,
                            "type": "string"
                          },
                          "city": {
                            "type": "string"
                          },
                          "province": {
                            "type": "string"
                          },
                          "postalCode": {
                            "type": "string"
                          },
                          "country": {
                            "type": "string"
                          },
                          "timeZone": {
                            "type": "string"
                          },
                          "checkInTime": {
                            "type": "string"
                          },
                          "checkOutTime": {
                            "type": "string"
                          },
                          "provincialRegistrationNumber": {
                            "nullable": true,
                            "type": "string"
                          },
                          "businessLicenceNumber": {
                            "nullable": true,
                            "type": "string"
                          },
                          "archivedAt": {
                            "nullable": true,
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                          },
                          "accessInstructions": {
                            "nullable": true,
                            "type": "string"
                          },
                          "defaultCleanerPayCents": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "addressLine1",
                          "addressLine2",
                          "city",
                          "province",
                          "postalCode",
                          "country",
                          "timeZone",
                          "checkInTime",
                          "checkOutTime",
                          "provincialRegistrationNumber",
                          "businessLicenceNumber",
                          "archivedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": ["items", "nextCursor"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["properties"]
      },
      "post": {
        "operationId": "Properties_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "addressLine1": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "addressLine2": {
                    "default": null,
                    "nullable": true,
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "city": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "province": {
                    "default": "BC",
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2
                  },
                  "postalCode": {
                    "type": "string",
                    "pattern": "^[A-Z]\\d[A-Z] ?\\d[A-Z]\\d$"
                  },
                  "timeZone": {
                    "default": "America/Vancouver",
                    "type": "string"
                  },
                  "checkInTime": {
                    "default": "16:00",
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "checkOutTime": {
                    "default": "11:00",
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "provincialRegistrationNumber": {
                    "default": null,
                    "nullable": true,
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "businessLicenceNumber": {
                    "default": null,
                    "nullable": true,
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "accessInstructions": {
                    "default": null,
                    "nullable": true,
                    "type": "string",
                    "maxLength": 2000
                  },
                  "defaultCleanerPayCents": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": ["name", "addressLine1", "city", "postalCode"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "addressLine1": {
                      "type": "string"
                    },
                    "addressLine2": {
                      "nullable": true,
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "province": {
                      "type": "string"
                    },
                    "postalCode": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "timeZone": {
                      "type": "string"
                    },
                    "checkInTime": {
                      "type": "string"
                    },
                    "checkOutTime": {
                      "type": "string"
                    },
                    "provincialRegistrationNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessLicenceNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "archivedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "accessInstructions": {
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultCleanerPayCents": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "addressLine1",
                    "addressLine2",
                    "city",
                    "province",
                    "postalCode",
                    "country",
                    "timeZone",
                    "checkInTime",
                    "checkOutTime",
                    "provincialRegistrationNumber",
                    "businessLicenceNumber",
                    "archivedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["properties"]
      }
    },
    "/v1/properties/{id}": {
      "get": {
        "operationId": "Properties_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "addressLine1": {
                      "type": "string"
                    },
                    "addressLine2": {
                      "nullable": true,
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "province": {
                      "type": "string"
                    },
                    "postalCode": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "timeZone": {
                      "type": "string"
                    },
                    "checkInTime": {
                      "type": "string"
                    },
                    "checkOutTime": {
                      "type": "string"
                    },
                    "provincialRegistrationNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessLicenceNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "archivedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "accessInstructions": {
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultCleanerPayCents": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "addressLine1",
                    "addressLine2",
                    "city",
                    "province",
                    "postalCode",
                    "country",
                    "timeZone",
                    "checkInTime",
                    "checkOutTime",
                    "provincialRegistrationNumber",
                    "businessLicenceNumber",
                    "archivedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["properties"]
      },
      "patch": {
        "operationId": "Properties_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "addressLine1": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "addressLine2": {
                    "nullable": true,
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "city": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "province": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2
                  },
                  "postalCode": {
                    "type": "string",
                    "pattern": "^[A-Z]\\d[A-Z] ?\\d[A-Z]\\d$"
                  },
                  "timeZone": {
                    "type": "string"
                  },
                  "checkInTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "checkOutTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "provincialRegistrationNumber": {
                    "nullable": true,
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "businessLicenceNumber": {
                    "nullable": true,
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "accessInstructions": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 2000
                  },
                  "defaultCleanerPayCents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "addressLine1": {
                      "type": "string"
                    },
                    "addressLine2": {
                      "nullable": true,
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "province": {
                      "type": "string"
                    },
                    "postalCode": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "timeZone": {
                      "type": "string"
                    },
                    "checkInTime": {
                      "type": "string"
                    },
                    "checkOutTime": {
                      "type": "string"
                    },
                    "provincialRegistrationNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessLicenceNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "archivedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "accessInstructions": {
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultCleanerPayCents": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "addressLine1",
                    "addressLine2",
                    "city",
                    "province",
                    "postalCode",
                    "country",
                    "timeZone",
                    "checkInTime",
                    "checkOutTime",
                    "provincialRegistrationNumber",
                    "businessLicenceNumber",
                    "archivedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["properties"]
      }
    },
    "/v1/properties/{id}/archive": {
      "post": {
        "operationId": "Properties_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "addressLine1": {
                      "type": "string"
                    },
                    "addressLine2": {
                      "nullable": true,
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    },
                    "province": {
                      "type": "string"
                    },
                    "postalCode": {
                      "type": "string"
                    },
                    "country": {
                      "type": "string"
                    },
                    "timeZone": {
                      "type": "string"
                    },
                    "checkInTime": {
                      "type": "string"
                    },
                    "checkOutTime": {
                      "type": "string"
                    },
                    "provincialRegistrationNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "businessLicenceNumber": {
                      "nullable": true,
                      "type": "string"
                    },
                    "archivedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "accessInstructions": {
                      "nullable": true,
                      "type": "string"
                    },
                    "defaultCleanerPayCents": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "addressLine1",
                    "addressLine2",
                    "city",
                    "province",
                    "postalCode",
                    "country",
                    "timeZone",
                    "checkInTime",
                    "checkOutTime",
                    "provincialRegistrationNumber",
                    "businessLicenceNumber",
                    "archivedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["properties"]
      }
    },
...
  },
  "info": {
    "title": "TruHost API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ]
}